+++ lib/features/ai/state/settings/ai_config_by_type_controller.dart
