+++ lib/features/ai/state/settings/inference_provider_form_controller.dart
