+++ test/features/ai/state/inference_model_form_controller_dirty_state_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/features/ai/model/ai_config.dart';
 import 'package:lotti/features/ai/repository/ai_config_repository.dart';
-import 'package:lotti/features/ai/state/inference_model_form_controller.dart';
+import 'package:lotti/features/ai/state/settings/inference_model_form_controller.dart';
 import 'package:mocktail/mocktail.dart';
 
 // Mocks
